requirements.txt
pytest-html==3.1.1
allure-pytest==2.10.0
PyMySQL==1.0.2
pytest==7.1.2
requests==2.28.1
selenium==4.20.0
WooCommerce==3.0.0
allure-pytest==2.10.0
python-dateutil
